% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{roughingit}
\alias{roughingit}
\title{"Roughing It"}
\format{A character vector with 13981 elements}
\source{
  \url{http://www.gutenberg.org/files/3177/3177.txt}
}
\usage{
  roughingit
}
\description{
  A dataset containing the full text of "Roughing It" by Mark Twain. 
  The UTF-8 plain text was sourced from the gutenberg.org website. 
  Each line represents a paragraph in the original text.
}
\keyword{datasets}
